+++ airplayvideo/service/model.py
@@ -11,7 +11,7 @@ import tempfile
 from urllib.parse import urlsplit, urlunsplit, parse_qs, urlencode
 import uuid
 
-VERSION = "0.1.4"
+VERSION = "0.1.5"
 IDENTIFIER = re.compile(r"[a-f0-9]{32}\Z")